LayoutTests/js/typedarray-findIndex-expected.txt

This test checks the behavior of the TypedArray.prototype.findIndex function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.findIndex.length is 1
PASS Int32Array.prototype.findIndex.name is 'findIndex'
PASS isSameFunctionForEachTypedArrayPrototype('findIndex') is true
PASS testPrototypeReceivesArray('findIndex', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Single Argument Testing
PASS testPrototypeFunction('findIndex', '(keepEven)', [12, 5, 8, 13, 44], 0) is true
PASS testPrototypeFunction('findIndex', '(keepEven)', [11, 13, 17, 13, 22], 4) is true
PASS testPrototypeFunction('findIndex', '(keepEven)', [11, 13, 17, 13, 11], -1) is true

2.0 Two Argument Testing
PASS testPrototypeFunction('findIndex', '(keepEven, thisValue)', [11, 23, 11, 1, 44], 3) is true

3.0 Array Element Changing
PASS testPrototypeFunction('findIndex', '(keepEvenAndChange)', [11, 15, 3, 12, 44], -1, [5, 5, 5, 5, 5]) is true

4.0 Exception Test
PASS testPrototypeFunction('findIndex', '(isBigEnoughAndException)', [12, 15, 10, 13, 44], 0) is true
PASS testPrototypeFunction('findIndex', '(isBigEnoughAndException)', [9, 15, 10, 13, 44], false) threw exception exception from function.

5.0 Wrong Type for Callback Test
PASS testPrototypeFunction('findIndex', '(8)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.findIndex callback must be a function.
PASS testPrototypeFunction('findIndex', '("wrong")', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.findIndex callback must be a function.
PASS testPrototypeFunction('findIndex', '(new Object())', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.findIndex callback must be a function.
PASS testPrototypeFunction('findIndex', '(null)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.findIndex callback must be a function.
PASS testPrototypeFunction('findIndex', '()', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.findIndex callback must be a function.

PASS successfullyParsed is true

TEST COMPLETE
